.history/sa/requirements_20250619232000.txt
# Install necessary packages (GPU version with CUDA 12.4)
# GPU 버전 PyTorch 설치 (CUDA 12.4)
--extra-index-url https://download.pytorch.org/whl/cu124
torch
torchvision

regex
pandas
numpy
tqdm
openpyxl
soynlp
jieba
mecab_ko
mecab-ko-msvc
mecab-ko-dic-msvc
prompt_toolkit

# 추가 필요
transformers>=4.21.0
FlagEmbedding>=1.2.0
sentence-transformers>=2.2.0
openai>=1.0.0
cohere>=4.0.0

FlagEmbedding>=1.2.0
soynlp>=0.0.493
numpy>=1.21.0
pandas>=1.3.0
openpyxl>=3.0.0
regex>=2021.8.3
torch>=1.9.0